+++ src/interest/presentation/interest.controller.ts
@@ -1,12 +1,12 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
 import { ApiBearerAuth, ApiExtraModels, ApiTags } from '@nestjs/swagger';
-import { AllInterestResDto, InterestResDto } from './dto/res.dto';
-import { Roles } from '../common/decorater/roles.decorator';
-import { Role } from '../user/enums/role.enum';
-import { InterestService } from './interest.service';
-import { InterestReqDto } from './dto/req.dto';
-import { ApiGetResponse } from '../common/decorater/swagger.decorator';
-import { Public } from '../common/decorater/public.decorator';
+import { AllInterestResDto, InterestResDto } from './res.dto';
+import { Roles } from '../../common/decorater/roles.decorator';
+import { Role } from '../../user/enums/role.enum';
+import { InterestService } from '../application/interest.service';
+import { InterestReqDto } from './req.dto';
+import { ApiGetResponse } from '../../common/decorater/swagger.decorator';
+import { Public } from '../../common/decorater/public.decorator';
 
 @Roles(Role.USER)
 @ApiTags('Interest')
