+++ src/subscriber/entity.subscriber.ts
@@ -2,7 +2,7 @@ import { Injectable, OnApplicationBootstrap } from '@nestjs/common';
 import { DataSource } from 'typeorm';
 import { RoleEntity } from '../user/infrastructure/entity/roles.entity';
 import { Role } from '../user/enums/role.enum';
-import { Interest } from '../interest/entity/interest.entity';
+import { Interest } from '../interest/infrastructure/entity/interest.entity';
 import { TechStack } from '../tech-stack/entity/tech-stack.entity';
 
 @Injectable()
